refactor: add nacl library in js
